compare as versions for compatibility
