+++ packages/eslint-plugin-ecmascript-compat/lib/compatibility.js
@@ -1,4 +1,5 @@
 /* eslint-disable camelcase, no-underscore-dangle */
+const compareVersions = require('./compareVersions');
 
 function unsupportedFeatures(features, targets) {
   return features.filter((feature) => !isFeatureSupportedByTargets(feature, targets));
@@ -30,7 +31,7 @@ function isCompatFeatureSupportedByTarget(compatFeature, target) {
     return true;
   }
 
-  return !support.isNone && target.version >= versionAdded;
+  return !support.isNone && compareVersions(target.version, versionAdded) >= 0;
 }
 
 function getSimpleSupportStatement(compatFeature, target) {
