+++ packages/eslint-plugin-ecmascript-compat/lib/targetRuntimes.js
@@ -1,6 +1,7 @@
 const browserslist = require('browserslist');
 const _ = require('lodash');
 const compatData = require('@mdn/browser-compat-data');
+const compareVersions = require('./compareVersions');
 
 module.exports = function targetRuntimes() {
   // ['chrome 50', ...]
@@ -64,20 +65,4 @@ function mapFamilyName(browserslistName) {
 
 function simplifyVersion(version) {
   return version.includes('-') ? version.split('-')[0] : version;
-}
-
-function compareVersions(a, b) {
-  const aParts = a.split('.').map(Number);
-  const bParts = b.split('.').map(Number);
-
-  for (let i = 0; i < Math.max(aParts.length, bParts.length); i++) {
-    if (aParts[i] === bParts[i]) {
-      // eslint-disable-next-line no-continue
-      continue;
-    }
-
-    return aParts[i] < bParts[i] ? -1 : 1;
-  }
-
-  return 0;
 }
